% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_events.R
\name{calc_centre}
\alias{calc_centre}
\title{Calculate population centre}
\usage{
calc_centre(.x, .method = "density")
}
\arguments{
\item{.x}{A numerical vector.}

\item{.method}{A character giving the method to use. Currently only density is available.}
}
\value{
A numeric
}
\description{
Calculate population centre
}
\examples{
\dontrun{
calc_centre(.x = rnorm(100))
}
}
\keyword{internal}
