% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_events.R
\name{calc_dist_to_centre}
\alias{calc_dist_to_centre}
\title{Calculate euclidean distance between two points.}
\usage{
calc_dist_to_centre(.x, .c)
}
\arguments{
\item{.x}{A numerical vector with coordinates to a point.}

\item{.c}{A numerical vector with the coordinates to the centre.}
}
\value{
A numerical vector with the euclidean distance between the two points.
}
\description{
Calculate euclidean distance between two points.
}
\note{
This function does mot make use of the base function dist, as that
dist-function is about twice as slow as the implementation here.
}
\examples{
\dontrun{
calc_dist_to_centre(.x = c(10, 15), .c = c(1, 2))
}
}
\keyword{internal}
