% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dist_chebyshev}
\alias{dist_chebyshev}
\title{Chebyshev distance}
\usage{
dist_chebyshev(x, diag = FALSE, upper = FALSE)
}
\arguments{
\item{x}{a numeric matrix or data frame.}

\item{diag}{logical value indicating whether the diagonal of the distance
matrix should be printed by \code{print.dist}.}

\item{upper}{logical value indicating whether the upper triangle of the
distance matrix should be printed by \code{print.dist} .}
}
\value{
Chebyshev distance returns an object of class "\code{dist}".
}
\description{
Chebyshev distance
}
\examples{

x <- matrix(rnorm(100), nrow = 5)
dist_chebyshev(x)

}
