% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyte_mfi.R
\name{harmonic_mean}
\alias{harmonic_mean}
\title{Calculate harmonic mean}
\usage{
harmonic_mean(.x)
}
\arguments{
\item{.x}{A numeric to take the harmonic mean of}
}
\value{
A single numeric value
}
\description{
Calculate harmonic mean
}
\section{Mean calculation}{

The calculation of the harmonic mean is as follows:
\deqn{
 n\frac{1}{\sum_{i=1}^n \frac{1}{x_i}}
}{length(x)/sum(1/x)}

\code{NA}s are removed before calculation
}

\examples{
\dontrun{
harmonic_mean(runif(10))
}
}
\keyword{internal}
