% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_analyte.R
\name{ident_bead_pop}
\alias{ident_bead_pop}
\title{Identify bead populations}
\usage{
ident_bead_pop(.analytes, .column_name = NULL, .cluster = NULL, .call_args,
  .data)
}
\arguments{
\item{.analytes}{A vector or list giving the IDs of the analytes.}

\item{.column_name}{A character giving the name of the column to subset by.}

\item{.cluster}{A character of the length of one giving the element to subset
by.}

\item{.call_args}{A list giving the parameters passed on to \code{identify_analyte()}.}

\item{.data}{A tidy data.frame.}
}
\value{
A data.frame.
}
\description{
This is a convenience function which allows to subset the data before calling
\code{\link[=identify_analyte]{identify_analyte()}}. The data is subset only if \code{.column_name} and
\code{.cluster} are given. Otherwise, the function is identical to calling
\code{\link[=identify_analyte]{identify_analyte()}} directly.
}
\details{
Convenience function to identify analytes in a subset
}
\examples{
x <- "a"
}
\keyword{internal}
