% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lplex}
\alias{lplex}
\title{LEGENDplex example data}
\format{A list with 18 elements. Each element is a data.frame about 5000 rows
and 4 columns (the exact number varies a little due to the data
acquisition):
\describe{
\item{FSC-A}{The forward scatter parameter}
\item{SSC-A}{The side scatter parameter}
\item{FL6-H}{Intensity in the FL6 channel}
\item{FL2-H}{Intensity in the FL2 channel}
}

The list contains 8 standard samples in duplicates, and one serum sample,
also in duplicate. The names of each element have the format K3 (internal
panel shorthand), C:num: for standards and S:num: for serum sample, and a
number indicating the replicate (1 or 2).}
\source{
Ulrik Stervbo, 2016, Unpublished
}
\usage{
data("lplex")
}
\description{
Data from a "Human Growth Factor Panel (13-plex)" LEGENDplex experiment, with
8 controls and 4 human serum samples, all in duplicates. The beads were
measured on a CytoFLEX cytometer, and the fcs-files were processed using
\code{\link[=read_fcs]{read_fcs()}}, with default settings.
}
\keyword{datasets}
