% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pad_matrix}
\alias{pad_matrix}
\title{Pad a matrix}
\usage{
pad_matrix(x, pad_size = 1L, pad = 0L)
}
\arguments{
\item{x}{The matrix to pad.}

\item{pad_size}{An integer giving the number of columns/rows to pad with.}

\item{pad}{A single value giving the padding.}
}
\value{
A matrix
}
\description{
Pad a matrix
}
\examples{

x <- matrix(c(1:9), nrow = 3, ncol = 3)
beadplexr:::pad_matrix(x)

x <- matrix(c(1:4), nrow = 2, ncol = 2)
beadplexr:::pad_matrix(x)
beadplexr:::pad_matrix(x, pad_size = 2L)
beadplexr:::pad_matrix(x, pad = NA)
beadplexr:::pad_matrix(x, pad = "xx")

x <- matrix(c(1:6), nrow = 3, ncol = 2)
beadplexr:::pad_matrix(x)

}
\keyword{internal}
