% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raise_deprecated.R
\name{raise_deprecated}
\alias{raise_deprecated}
\title{Warn about deprecated parameters or functions}
\usage{
raise_deprecated(old, new, caller = NULL)
}
\arguments{
\item{old}{A character giving the old function or parameter.}

\item{new}{A character giving the replacement function or parameter.}

\item{caller}{A character vector giving the function with the deprecated
parameter.}
}
\value{
Nothing. Just raise a warning
}
\description{
Warn about deprecated parameters or functions
}
\examples{
beadplexr:::raise_deprecated("XX", "YY")
beadplexr:::raise_deprecated(old = "XX", new = "YY", caller = "ZZZ")
beadplexr:::raise_deprecated(old = "XX", new = "YY", caller = "ZZZ()")

}
\keyword{internal}
